:truck: move client stuff into client subpackage
