+++ pyUSIrest/__init__.py
@@ -8,7 +8,6 @@ __version__ = '0.3.0.dev0'
 
 from . import auth
 from . import client
-from . import document
 from . import settings
 
-__all__ = ["auth", "client", "document", "settings"]
+__all__ = ["auth", "client", "settings"]
